topology: Define HDA_DAI_CONFIG to configure HDA backend links
this patch define macro HDA_DAI_CONFIG(idx, link_id, name) for user to
define configurations for HDA Backend links.

Pleae note, there is not HW configurations defined for HDA BE links
because current HW configurations is more suitable for SSP BE links.
But the data tuple will carry the HDA type info for the link for kernel
driver to check the link type.

Signed-off-by: Mengdong Lin <mengdong.lin@linux.intel.com>