[RFC]topology: fix kcontrol name by topology
the kcontrol item name doesn't follow the SOURCE:DIRECTION:FUNCTION, and
can't seperate playback and capture when display in alsamixer.
This fix is to move the pipeline id from the end of the name to the
head, this will make alsamixer to serperate playback and capture when
press F3 F4.

Signed-off-by: Zhu Yingjiang <yingjiang.zhu@linux.intel.com>